% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disj_Clst_SG.R
\name{Disj_Clst_SG}
\alias{Disj_Clst_SG}
\title{Convert an SG's IBD structure into a disjoint clusters}
\usage{
Disj_Clst_SG(SG, IBDClusters)
}
\arguments{
\item{SG}{Concerned SG}

\item{IBDClusters}{The output of Dash, which is used to build SG, shows which clusters share the same haplotype (IBD)}
}
\value{
SG with Disjoint clusters
}
\description{
Use graph theory to convert an SG's IBD structure into a disjoint clusters
}
\examples{
Disj_Clst_SG(SG, IBDClusters)



}
