% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disj_Clst_SG_Bigger.R
\name{Disj_Clst_SG_Bigger}
\alias{Disj_Clst_SG_Bigger}
\title{Convert an SG's IBD structure into a disjoint clusters}
\usage{
Disj_Clst_SG_Bigger(SG, IBDClusters)
}
\arguments{
\item{SG}{Concerned SG}

\item{IBDClusters}{The output of Dash, which is used to build SG, shows which clusters share the same haplotype (IBD)}
}
\value{
SG with Disjoint clusters
}
\description{
This code first sorts the groups by size in descending order. Then it checks each group for overlap with the groups already. When there are overlaps, the largest group is kept.
}
\examples{
Disj_Clst_SG_Bigger(SG, IBDClusters)



}
