% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_colors.R
\name{generate_random_colors}
\alias{generate_random_colors}
\title{Generate a random color that is not too dark}
\usage{
generate_random_colors(N)
}
\arguments{
\item{N}{: Number of code color to generate}
}
\value{
Color code of N colors
}
\description{
Generate a random color that is not too dark
}
\examples{
generate_random_colors(4)

}
