% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.MergeSG.R
\name{gr.MergeSG}
\alias{gr.MergeSG}
\title{Merging two SGs Grange}
\usage{
gr.MergeSG(gr.SG1, gr.SG2)
}
\arguments{
\item{gr.SG1}{Grange of the first SG}

\item{gr.SG2}{Grange of the second SG}
}
\value{
Merged Grange
}
\description{
Merging two SGs Grange
}
\examples{
MergeSG(gr.SG1,gr.SG2)
MergeSG(gr.Sliding_SG,gr.SG)


}
