% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.Sliding_SG.R
\name{gr.Sliding_SG}
\alias{gr.Sliding_SG}
\title{@title Creating a Grange of Sliding_SG}
\usage{
gr.Sliding_SG(gr.SG, overlapSS = 10)
}
\arguments{
\item{gr.SG}{:  Grange  of concerned SG}

\item{overlapSS}{: Overlap between SG and Sliding_SG in Kbp : default value is 10 Kbp}
}
\value{
: Grange of Sliding_SG
}
\description{
@description  Creating a Grange of Sliding_SG associate to a specific SG (to be specified). Length of Sliding_SG is twice the overlapSS (in Kb). For creating a more specific Sliding_SG gr.Next_SG function  give more flexibility
}
\examples{
Sliding_SG(gr.SG)
Sliding_SG(gr.SG=gr.SG, overlapSS=10)
Sliding_SG(gr.SG, 10)

}
