% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBD_structure.R
\name{IBD_structure}
\alias{IBD_structure}
\title{identify the IBD structure of an SG (or Sliding-SG)}
\usage{
IBD_structure(gr.SG, IBDClusters, overlap = 0.5)
}
\arguments{
\item{gr.SG}{Grange of concerned SG}

\item{IBDClusters}{Dash output in data frame format}

\item{overlap}{the cutoff point at which the cluster is regarded as part of the structure}
}
\value{
Clusters of IBD structure
}
\description{
identify the IBD structure of an SG (or Sliding-SG)
}
\examples{
IBD_structure(gr.SG, IBDClusters)
IBD_structure(gr.SG=gr.SG, IBDClusters=df, overlap=0.5)
IBD_structure(gr.SG=gr.Sliding_SG, IBDClusters=df, overlap=0.5)

}
