% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IBDClusters}
\alias{IBDClusters}
\title{A Dash output example that can be used}
\format{
\subsection{\code{IBDClusters}}{

A data frame with 27 rows and 13 columns:
\describe{

\item{Column 1}{Cluster identifier}
\item{Column 2}{Cluster start position}
\item{Column 3}{Cluster end position}
\item{Column 4}{Maximum IBD match start position}
\item{Column 5}{Minimum IBD match end position}
\item{Column 6}{Family and Sample ID for cluster carriers ...}
...
}
}
}
\source{
{Created in-house to serve as example}
data(IBDClusters)
}
\usage{
IBDClusters
}
\description{
This is the outcome of applying Dash to a chromosome 22 simulation chip.
}
\keyword{datasets}
