% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_same_IBDstructure.R
\name{is_same_IBDstructure}
\alias{is_same_IBDstructure}
\title{Compare two SGs' IBD structures}
\usage{
is_same_IBDstructure(SG1, SG2)
}
\arguments{
\item{SG1}{The first SG: IBD_structure is the second element in SG, which is a list of two elements.}

\item{SG2}{The second SG}
}
\value{
Return True if SG1 and SG2 have the same IBD_structure otherwise False
}
\description{
Compare two SGs' IBD structures
}
\examples{
is_same_IBDstructure(SG1, SG2)
}
