% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sall.R
\name{Sall}
\alias{Sall}
\title{Computing Sall for all SGs of chromosome}
\usage{
Sall(SGList, Naff = NULL)
}
\arguments{
\item{SGList}{list of SG with disjoints IBD clusters for a specific chromosome}

\item{Naff}{Number of affected individuals}
}
\value{
Dataframe with the value of Sall for each SG, Sall proposed by Whittemore and Halpernand and the Sall_WH per affected if Naff is gived in input
}
\description{
Computing Sall for all SGs of chromosome
}
\examples{
Sall(SGList)
Sall(SGList[[1]])
Sall(SGList[1:5])



}
