% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sall_SG.R
\name{Sall_SG}
\alias{Sall_SG}
\title{Computing Sall for a unique SG with disjoint clusters}
\usage{
Sall_SG(SG)
}
\arguments{
\item{SG}{: One SG example  SG=SGList[\link{1}]}
}
\value{
The value of Sall for this SG, a numeric value
}
\description{
Computing Sall for a unique SG with disjoint clusters
}
\examples{
 Sall_SG(SG=SGList[[1]])


}
