% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smsg.R
\name{Smsg}
\alias{Smsg}
\title{Function to compute Smsg}
\usage{
Smsg(SGList, Naff = NULL)
}
\arguments{
\item{SGList}{List of SG with disjoints IBD clusters for a given chromosome}

\item{Naff}{Number of affected individuals}
}
\value{
Dataframe with the value of Smsg for each SG and, in the event that Naff is provided as input, the Smsg per affected also is calculated
}
\description{
Function to compute Smsg
}
\examples{
Smsg(SGList)
Smsg(SGList[[1]])
Smsg(SGList[1:5])



}
