% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBM.R
\name{TBM}
\alias{TBM}
\title{Function based on the tunnel-boring machine (TBM) principle to creat a list of SGs}
\usage{
TBM(
  IBDClusters,
  SGList = list(),
  chr = NULL,
  N = NULL,
  overlapSS = 10,
  overlapNS = 0,
  overlapI = 0.5,
  MinL = 400
)
}
\arguments{
\item{IBDClusters}{IBD clusters (formatted as a data frame from Dash output)}

\item{SGList}{list of SGs}

\item{chr}{Concerned chromosome}

\item{N}{Number of iteration, The NULL value is used to scan the entire chromosome.}

\item{overlapSS}{Overlap between Sliding-SG and SG grange; 10Kbp is the default value}

\item{overlapNS}{Overlap between the SG grange and the Next-SG grange; by default, this is set to 0}

\item{overlapI}{the threshold at which the cluster is considered to be included in the IBD structure}

\item{MinL}{The SG Grange's minimum length in Kbp is 400 Kbp by default}
}
\value{
list of SGs
}
\description{
Based on TBM_SG function to scan all chromosome
}
\examples{
TBM(SGList, IBDClusters)
TBM(SGList, IBDClusters, overlapSS=10, overlapNS=0, overlapI=0.5, MinL=400)


}
