% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBM_SG.R
\name{TBM_SG}
\alias{TBM_SG}
\title{Function based on the tunnel-boring machine (TBM) principle to create one SG;}
\usage{
TBM_SG(
  SG,
  IBDClusters,
  overlapSS = 10,
  overlapNS = 0,
  overlapI = 0.5,
  MinL = 400
)
}
\arguments{
\item{SG}{Concerned SG}

\item{IBDClusters}{IBD clusters (formatted as a data frame from Dash output)}

\item{overlapSS}{Overlap between Sliding-SG and SG grange; 10Kbp is the default value}

\item{overlapNS}{Overlap between the SG grange and the Next-SG grange; by default, this is set to 0}

\item{overlapI}{the threshold at which the cluster is considered to be included in the IBD structure}

\item{MinL}{The SG Grange's minimum length in Kbp is 400 Kbp by default}
}
\value{
A list of two elements; the Grange of SG and IBD structure
}
\description{
Function based on the tunnel-boring machine (TBM) principle to create one SG;
primarily designed to expand the existing SG if its IBD structure like that of a sliding-SG;
else it produces a new nearby SG
}
\examples{
TBM_SG(SG, IBDClusters)
TBM_SG(SG, IBDClusters, overlapSS=10, overlapNS=0, overlapI=0.5, MinL=400)




}
