% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBMD.R
\name{TBMD}
\alias{TBMD}
\title{convert the non-disjoint clusters in the SGlist output of TBM into the same output with segregated clusters.}
\usage{
TBMD(SGList, IBDClusters, disjM = "Merge", remove = TRUE)
}
\arguments{
\item{SGList}{The TBM function's output or any list with the same structure}

\item{IBDClusters}{Data frame format of dash output}

\item{disjM}{To construct disjoint clusters, you have two methods available. If you want to merge overlapping clusters, set disjM to "Merge". Alternatively, if you prefer to select the cluster with the most shared haplotype when overlaps occur, set disjM to "Bigger". Choose the method that best suits your data analysis needs.}

\item{remove}{Argument to exclude from the IBD structure any SG that does not have any clusters}
}
\value{
List of SG with disjoints clusters in IBD structure
}
\description{
convert the non-disjoint clusters in the SGlist output of TBM into the same output with segregated clusters.
}
\examples{
TBMD(SGList, IBDClusters)
TBMD(SGList, IBDClusters, disjM="Merge", remove=TRUE)
TBMD(SGList, IBDClusters, disjM="Bigger", remove=TRUE)



}
